requirements.txt
Django==3.2.13
asgiref==3.4.1
dj-database-url==0.5.0
gunicorn==20.1.0
pycodestyle==2.8.0
pytz==2021.3
sqlparse==0.4.2
toml==0.10.2
whitenoise==5.3.0
django-environ==0.8.1
aiosmtpd==1.4.2
black==22.6.0
cryptography==37.0.4